LSR-860 Same code repeated again. Pull it up too
